pyproject.toml
[tool.poetry]
name = "ai-workflow"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ability", from = "src"}, {include = "core", from = "src"}, {include = "operator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
pydantic = "^2.11.5"
aiohttp = "^3.12.12"
fastmcp = "^2.8.0"
langchain-mcp-adapters = "^0.1.7"
langchain = "^0.3.25"
langchain-openai = "^0.3.22"
langchain-anthropic = "^0.3.15"
langgraph = "^0.4.8"
requests = "^2.32.4"
pandas-ta = "^0.3.14b0"
feedparser = "^6.0.11"
textblob = "^0.19.0"
asyncpraw = "^7.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"